// src/redux/types/mapping.type.ts
/**
 * Copyright (c) 2023, RTE (http://www.rte-france.com)
 * This Source Code Form is subject to the terms of the Mozilla Public
 * License, v. 2.0. If a copy of the MPL was not distributed with this
 * file, You can obtain one at http://mozilla.org/MPL/2.0/.
 */

export interface BasicProperty {
    name: string;
    value: string;
    type: string; // TODO must be PropertyType enum
}

export interface Automaton {
    family: string;
    model: string;
    setGroup: string;
    properties: BasicProperty[];
}